Review notes from Daniel

+++ robot/include/anki/cozmo/robot/logging.h
@@ -106,7 +106,7 @@ namespace Anki {
       { Anki::Cozmo::RobotInterface::SendLog(Anki::Cozmo::RobotInterface::ANKI_LOG_LEVEL_DEBUG, nameId, fmtId, nargs, CASTx(nargs, __VA_ARGS__)); }
       
       #define AnkiDebugPeriodic(num_calls_between_prints, nameId, nameString, fmtId, fmtString, nargs, ...) \
-      {   static u32 cnt = num_calls_between_prints; \
+      {   static u16 cnt = num_calls_between_prints; \
           if (cnt++ >= num_calls_between_prints) { \
             Anki::Cozmo::RobotInterface::SendLog(Anki::Cozmo::RobotInterface::ANKI_LOG_LEVEL_DEBUG, nameId, fmtId, nargs, CASTx(nargs, __VA_ARGS__)); \
             cnt = 0; \
